Prebid Auctions
The Prebid Auction is introduced, making it possible to manage a series of auction events. Previously Prebid handled only one auction at a time. With this PR the `pbjs.adUnits` and `pbjs.bidderSettings` are moved to an `auction` instance. Auctions can be run concurrently, so for example if different groupings of ad placements have different configurations, auctions for each placement group can both be requested at page load and run asynchronously. Be aware that if multiple auctions are run concurrently for the same placement the last auction winner creative will replace any previous winner.